Expand repo scope for cross-component architecture docs
I feel we need a home for architecture and design documentation for
issues that span multiple components. The schema is one example of a
detail that needs to be coordinated across components, but there are
others, particularly with backend automation, and automation with
GitHub to assist in taxonomy development, triage, and review.

A follow-up to this change should be to rename the repo to a more
generic name that reflects an expanded scope of artifacts.

Part of issue #4

Signed-off-by: Russell Bryant <rbryant@redhat.com>